Fix the issues related with Omniscient and Twitter integration

+++ dragonfire/utilities.py
@@ -85,31 +85,32 @@ class TTA:
         if self.twitter:
             text = "@" + self.twitter_user + " " + message#.upper()
             text = (text[:TWITTER_CHAR_LIMIT]) if len(text) > TWITTER_CHAR_LIMIT else text
-            if len(self.command) > 1:
-                if self.command[0] == "sensible-browser":
-                    reduction = len(text + " " + self.command[1]) - TWITTER_CHAR_LIMIT
-                    if reduction < 1:
-                        reduction = None
-                    text = text[:reduction] + " " + self.command[1]
-                    page = metadata_parser.MetadataParser(url=self.command[1])
-                    img_url = page.get_metadata_link('image')
-                    if img_url:
-                        response = urllib.request.urlopen(img_url)
-                        img_data = response.read()
-                        img_extension = mimetypes.guess_extension(response.headers['content-type'])
-                        filename = "/tmp/tmp" + uuid.uuid4().hex + img_extension
-                        with open(filename, 'wb') as f:
-                            f.write(img_data)
+            if self.command:
+                if len(self.command) > 1:
+                    if self.command[0] == "sensible-browser":
+                        reduction = len(text + " " + self.command[1]) - TWITTER_CHAR_LIMIT
+                        if reduction < 1:
+                            reduction = None
+                        text = text[:reduction] + " " + self.command[1]
+                        page = metadata_parser.MetadataParser(url=self.command[1])
+                        img_url = page.get_metadata_link('image')
+                        if img_url:
+                            response = urllib.request.urlopen(img_url)
+                            img_data = response.read()
+                            img_extension = mimetypes.guess_extension(response.headers['content-type'])
+                            filename = "/tmp/tmp" + uuid.uuid4().hex + img_extension
+                            with open(filename, 'wb') as f:
+                                f.write(img_data)
 
-                        try:
-                            self.twitter_api.update_with_media(filename, text)
-                            if randint(1,3) == 1:
-                                self.twitter_api.create_friendship(self.twitter_user, follow=True)
-                        except TweepError as e:
-                            print("Warning: " + e.response.text)
-                        finally:
-                            os.remove(filename)
-                        return True
+                            try:
+                                self.twitter_api.update_with_media(filename, text)
+                                if randint(1,3) == 1:
+                                    self.twitter_api.create_friendship(self.twitter_user, follow=True)
+                            except TweepError as e:
+                                print("Warning: " + e.response.text)
+                            finally:
+                                os.remove(filename)
+                            return True
             try:
                 self.twitter_api.update_status(text)
                 if randint(1,3) == 1:
